refactor: update import paths and enhance library handling in Wyoming Voiceprint service
