--- web_ui/src/lib/state/index.ts
@@ -1,6 +1,0 @@
-export { VoiceprintProvider } from "./voiceprint-provider";
-export { useGetLibraries } from "./use-get-libraries";
-export { useCurrentLibrary } from "./use-current-library";
-export { useCreateLibrary } from "./use-create-library";
-export { useVoiceprintContext } from "./use-voiceprint-context";
-export type { VoiceprintContextValue } from "./voiceprint-context";
